Implement basic YouTube audio, formatting
Bot can join and have basic musicbot functionality, and single-quote
strings now use double-quote for consistency.

Improvement could be made in audio quality, consistency, and processing,
as the bot takes a long time to process a youtube playlist given.

Note: the previous play function, which played from local files,
is replaced with the yt function, which inherits the name, considering how
the bot will be used.