Isolate private token
The private bot token was previously stored as a string in the main file.
For obvious reasons this is a security flaw. The token is now stored in
a separate Python file, retrieved by an import. The private directory
housing the file has been added to a newly created gitignore,
along with other files.

The token previously stored has been made obsolete with a newly generated
token, for safety.